Create README.md
DVD Rental Store Application After start of the program user should see the possible options and actions that can be done

User should be able to see current offer: which movies and how many copies are available in the store After providing client id user should see all the rentals connected to this client. Display the results in two sections: active rentals (those not yet returned) and historic rentals -There should be an option to create a new rental, using client id and copy id with current date as date of rental Of course there should also be an option for registering a return of a copy Allow the creation of a new user Another option should be for creating a new movie in our store. Creating a new movie would mean, that a single copy of that movie should also be created. Tip: transaction should be used here An owner of the DVD rental store would probably be interested in some statistics. Allow the option for displaying the number of rentals as well as total price of rented movie from date of users choice to now. Add the functionality of displaying rentals that are overdue - more than two weeks after date of rental they are still not returned. Your program should use at least one of the design patterns we've learned during laboratories